tidy,docs(ui): focus region logic
